refactor VTI with commander.js
