+++ vti/tsconfig.json
@@ -5,7 +5,5 @@
     "outDir": "./dist",
     "rootDir": "src"
   },
-  "include": [
-    "src"
-  ]
+  "include": ["src/**/*"]
 }
